BOL-62 refactor: group search uimodel 패키지 변경
