+++ presentation/src/main/java/com/yapp/bol/presentation/view/group/search/GroupSearchViewModel.kt
@@ -4,6 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import androidx.paging.*
 import com.yapp.bol.domain.usecase.group.SearchGroupByKeywordUseCase
+import com.yapp.bol.presentation.data.GroupSearchUiModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
